requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
cryptography==41.0.7
web3==6.11.3
requests==2.31.0
python-decouple==3.8
django-extensions==3.2.3
pillow==10.1.0
stripe==7.8.0
pycryptodome==3.19.0
django-filter==23.4
django-money==3.4.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
gunicorn==21.2.0
whitenoise==6.6.0 